Beta Fix - issue where example tokens were being removed

+++ Token.js
@@ -1493,7 +1493,7 @@ class Token {
 
 	place(animationDuration) {
 		try{
-			if(isNaN(parseFloat(this.options.left)) || isNaN(parseInt(this.options.top))){// prevent errors with NaN positioned tokens - delete them as catch all. 
+			if(!this.options.id.includes('exampleToken') && (isNaN(parseFloat(this.options.left)) || isNaN(parseInt(this.options.top)))){// prevent errors with NaN positioned tokens - delete them as catch all. 
 				this.options.deleteableByPlayers = true;
 				this.delete();
 				return;
